Move MBR to criterion
